pyproject.toml
[tool.poetry]
name = "algo_trader"
version = "0.1.0"
description = ""
authors = ["kbains09 <kiratbains09@gmail.com>"]
readme = "README.md"
packages = [{ include = "algo_trader", from = "src" }]
# ^ This tells Poetry to look inside `src/algo_trader`

[tool.poetry.dependencies]
python = ">=3.13"
pandas = "^2.3.1"
numpy = "^2.3.2"
scikit-learn = "^1.7.1"
yfinance = "^0.2.65"
ta = "^0.11.0"
statsmodels = "^0.14.5"
stable-baselines3 = "^2.7.0"
gym = "^0.26.2"
ta-lib = "^0.6.5"
xgboost = "^3.0.4"
joblib = "^1.5.1"
typer = "^0.16.0"
pyyaml = "^6.0.3"
pyarrow = "^22.0.0"
fastparquet = "^2024.11.0"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.1.1"
ipykernel = "^6.30.1"
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
